Port same_name test to new integration test style
